Add crud adm Paguese a la orden de,OP

+++ Convertidor/Data/EntitiesDestino/ADM/ADM_BENEFICIARIOS_OP.cs
@@ -6,9 +6,9 @@
         public int CODIGO_ORDEN_PAGO { get; set; }
         public int CODIGO_PROVEEDOR { get; set; }
         public int CODIGO_CONTACTO_PROVEEDOR { get; set; }
-        public float MONTO { get; set; }
-        public float MONTO_PAGADO { get; set; }
-        public float MONTO_ANULADO { get; set; }
+        public decimal MONTO { get; set; }
+        public decimal MONTO_PAGADO { get; set; }
+        public decimal MONTO_ANULADO { get; set; }
         public string EXTRA1 { get; set; } = string.Empty;
         public string EXTRA2 { get; set; } = string.Empty;
         public string EXTRA3 { get; set; } = string.Empty;
